Refactor GetEngineBinaryClient
Longhorn 4210
Longhorn 3198
